fix code quality issue

+++ src/execution/reservoir_sample.cpp
@@ -492,7 +492,7 @@ void ReservoirSamplePercentage::Finalize() {
 	    static_cast<double>(current_count) > sample_percentage * RESERVOIR_THRESHOLD || finished_samples.empty();
 	if (current_count > 0 && sampled_more_than_required) {
 		// create a new sample
-		auto new_sample_size = static_cast<idx_t>(round(sample_percentage * current_count));
+		auto new_sample_size = static_cast<idx_t>(round(sample_percentage * static_cast<double>(current_count)));
 		auto new_sample =
 		    make_uniq<ReservoirSample>(allocator, new_sample_size, base_reservoir_sample->random.NextRandomInteger());
 		while (true) {
